msm: scm-boot: move scm-boot driver to a new location in kernel
This patch moves scm-boot driver from mach-msm directory to
machine independent location in a kernel tree.

+++ drivers/soc/qcom/scm-boot.c
@@ -21,8 +21,8 @@
 int scm_set_boot_addr(phys_addr_t addr, unsigned int flags)
 {
 	struct {
-		u32 flags;
-		u32 addr;
+		unsigned int flags;
+		unsigned long addr;
 	} cmd;
 
 	cmd.addr = addr;
